#48142 Choose column is not working in Job Advance report

+++ QA/M4PL/M4PL.Entities/Job/JobAdvanceReport.cs
@@ -35,10 +35,10 @@ namespace M4PL.Entities.Job
 		public DateTime? JobDeliveryDateTimeActual { get; set; }
 		public string JobCustomerPurchaseOrder { get; set; }
 		public decimal? JobTotalCubes { get; set; }
-		public decimal? TotalParts { get; set; }
+		public int? TotalParts { get; set; }
 		public string JobNotes { get; set; }
 		public string JobCarrierContract { get; set; }
-		public decimal? TotalQuantity { get; set; }
+		public int? TotalQuantity { get; set; }
         public string JobSiteCode { get; set; }
 
     }
